refactor: change post objects directories
